move get repository functions to dependencies file
